major refactor.
